+++ LMS_CMS_Backend/LMS_CMS_BL/DTO/Accounting/FeesActivationGetDTO.cs
@@ -12,9 +12,9 @@ namespace LMS_CMS_BL.DTO.Accounting
     {
         [Key]
         public long FeeActivationID { get; set; }
-        public float Amount { get; set; }
+        public decimal Amount { get; set; }
         public float Discount { get; set; }
-        public float Net { get; set; }
+        public decimal Net { get; set; }
         public string Date { get; set; }
         public long FeeTypeID { get; set; }
         public string? FeeTypeName { get; set; }
